+++ core/state/statedb_hooked.go
@@ -289,6 +289,6 @@ func (s *hookedStateDB) Finalise(deleteEmptyObjects bool) *bal.ConstructionBlock
 	return s.inner.Finalise(deleteEmptyObjects)
 }
 
-func (s *hookedStateDB) SetTxContext(thash common.Hash, ti int, blockAccessIndex uint16) {
+func (s *hookedStateDB) SetTxContext(thash common.Hash, ti int, blockAccessIndex uint32) {
 	s.inner.SetTxContext(thash, ti, blockAccessIndex)
 }
