+++ core/vm/interface.go
@@ -99,5 +99,5 @@ type StateDB interface {
 
 	// Finalise must be invoked at the end of a transaction
 	Finalise(bool) *bal.ConstructionBlockAccessList
-	SetTxContext(thash common.Hash, ti int, blockAccessIndex uint16)
+	SetTxContext(thash common.Hash, ti int, blockAccessIndex uint32)
 }
